Minor fixes in Stack Monitor module

+++ Examples/VPTemplate/src/Util/StackMonitor/StackMonitor.h
@@ -7,7 +7,7 @@
 #define STACK_VALUE 0xDEADBEEF
 
 extern int32_t _size_of_stack;
-extern uint8_t _bottom_of_stack;
+extern int32_t _bottom_of_stack;
 
 int32_t getFreeBytes();
 int32_t getUsedBytes();
